test in UnicodeConfigTest verschoben
